Add files via upload
This C++ application prompts the user to enter an integer and then checks whether this number is a power of 2. It employs a loop that continuously divides the number by 2, checking at each step if the result is even. If the division results in a number other than 2 or an odd value, the loop stops. By the end, if the number is reduced to 2, it confirms that the original number is indeed a power of 2. The program then provides the user with the verdict.